[1.2] Added Widget-0.5.3 and language selector
Hello everyone!

This update contains the newest version of Widget, version 0.5.3. You can download the zip through the repository or the website. You can now type text in the app and view this at a widget, which are now in 6 colors!
This update also brings a selector to choose between languages, which is at the moment Dutch and English.

Have fun with this update!
Ryan